Fix clippy error casting same size

+++ core/io/windows.rs
@@ -119,7 +119,7 @@ impl File for WindowsFile {
 
     fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         let mut file = self.file.borrow_mut();
-        file.set_len(len).map_err(LimboError::IOError)?;
+        file.set_len(len as u64).map_err(LimboError::IOError)?;
         c.complete(0);
         Ok(())
     }
